Update the checksum of the source files when they are edited.
The checksum is updated only when the edited file is not the `PKGBUILD` or an `.install` file.

For example:

	trizen -S linpack

It contains the file "linpack.conf". When this file is edited, the checksum are updated by running the `updpkgsums` tool from Pacman.

Thanks to Matthias for reporting this issue by email.