Add option to force color output
The parameter or config option 'forcecolors' forces color output even
when not writing to STDOUT. It takes precedence over 'nocolors'.
Implementing this was a little bit hacky due to the fact that the config
file is evaluated after the commandline parameters.
The $colors variable used can hold the following 4 values:
0 -> force colors off
1 -> colors off because of no STDOUT
2 -> default colors on
3 -> force color on