- Load all the dependencies of `trizen` before performing `-Su`.
If `perl` gets updated along with a Perl module that has some parts written in C, then the
module will fail to load later on, since `trizen` was executed with the older verison of `perl`.
Therefore, we load all the critical modules before performing `-Su`, to prevent a potential crash in the future. (fixes #206)