- Check if the VCS dependencies are out-of-date, during `--devel`.
When a package is now updated or installed with `--devel`, trizen will check if its VCS installed dependencies (git only, for now) are all up-to-date.

If not, then the out-of-date VCS dependencies will be updated first. After which the main package will be built and installed.

This helps with updating "intel-media-driver-git", which often requires the latest commit of "libva-git" in order to build successfully.